Add brave_vpn_region_data_helper to avoid internal rules
