+++ components/brave_vpn/browser/brave_vpn_service_helper.h
@@ -22,18 +22,6 @@ class Value;
 }  // namespace base
 
 namespace brave_vpn {
-mojom::RegionPtr GetRegionPtrWithNameFromRegionList(
-    const std::string& name,
-    const std::vector<mojom::RegionPtr>& region_list);
-base::Value::Dict GetValueFromRegionWithoutCity(const mojom::RegionPtr& region);
-base::Value::Dict GetValueFromRegion(const mojom::RegionPtr& region);
-bool IsValidRegionValue(const base::Value::Dict& value);
-mojom::RegionPtr GetRegionFromValueWithoutCity(const base::Value::Dict& value);
-mojom::RegionPtr GetRegionFromValue(const base::Value::Dict& value);
-bool ValidateCachedRegionData(const base::Value::List& region_value);
-std::vector<mojom::RegionPtr> ParseRegionList(
-    const base::Value::List& region_list);
-
 // False if subscription is expired.
 bool IsValidCredentialSummary(const base::Value& summary);
 bool IsValidCredentialSummaryButNeedActivation(const base::Value& summary);
